scripts: validate-boot: make use of mk-flash-layout if present
The mount_ro function was modified to check /var/state/flash_layout.rc
then /sbin/mk-flash-layout if present. This allows automatic flash device
detection.

This program is specific in that it knows about a list of MBRs taken
from the command line. In order to keep backwards compatibility, the
list is still considered but all read operations are made on the MBR
detected.