scripts: mk-flash-layout: scan partitions only when booted from compliant bootloader
We don't try to scan or mount partitions anymore if "img=" is not found
on the kernel command line, because this means we're not on a compliant
system. However, the scan is still performed if "-s" is passed on the
script's command line, in order to allow easier system recovery from
maintenance images.