scripts: mk-flash-layout: accept 8064 kB as the lower limit for the SW part
Most flash block devices are formated at 63 sectors per track and 32 heads
per cylinder, or 0.984 MB. Having a minimum limit of 8192 kB forces to
allocate one more megabyte just to save the missing 1.6% that results from
the 63/64 ratio. Let's accept 8064 (8 cylinders) as the lower limit in
order to make it simpler to work with small config-only devices in VMs.