scripts/update-boot-image: don't try to unmount if mount failed
When the script is called from within the firmware, it cannot succeed
on /sbin/mount-flash because this one is not present in the firmware,
but it still tries to unmount, resulting in an error message being
displayed. The error is harmless as this is only done in order to
update the flash-layout.rc file after the upgrade, which is not used
there. Simply avoid setting the unmount flag on mount error.