scripts/update-boot-image: update the flash's journal upon image upgrade
Now we update the flash partition's journal when an image is updated or
when an attempt is made and fails (in which case the failure cause will
also be reported).

The journal will indicate the model and version of the image that was
uploaded, which image was about to be replaced, which image(s) are
removed if any, and where the image is finally installed.

One tricky aspect is that the script supports being called from the
firmware where commands are very limited. There's no date command nor
any tty command. The date is important so here we attempt to find it
using various ways (ls and tar emit a usable one after some processing).
The tty may be reported as "none" if undetermined. The product, model
and version will report "unknown" if run from the firmware. Otherwise
the rest of the code still works. An example of output looks like this :

2016-07-07 04:51:54 unknown unknown unknown none update-boot-image: mount: flash mounted R/W to write image version alb3100-8.0.4-8009 as backup
2016-07-07 04:51:54 unknown unknown unknown none update-boot-image: rm: removed previous backup image after mv succeeded
2016-07-07 04:51:54 unknown unknown unknown none update-boot-image: mv: New image installed as backup
2016/07/07 07:03:29 aloha alb3100 8.0.4 pts/0 update-boot-image: mount: flash mounted R/W to write image version alb3100-7.5.5-7537 as active
2016/07/07 07:03:29 aloha alb3100 8.0.4 pts/0 update-boot-image: rm: removed previous active image after mv succeeded
2016/07/07 07:03:29 aloha alb3100 8.0.4 pts/0 update-boot-image: mv: New image installed as active