scripts/update-boot-image: do not corupt flash-layout.rc when out of space
We are using a temporary file in /var/state, to compute the new flash-layout.
When /var/state if full, we cannot build this file correctly, but we still
leave an empty file.
Then we replace /var/state/flash-layout.rc with this empty file.

We now ensure that we jave enough space to create the new flash layout,
otherwise we leave it unchanged.